Moved types-specifying extensions to PHPStan\Type\Php
