+++ src/Type/Php/IsFloatFunctionTypeSpecifyingExtension.php
@@ -1,18 +1,18 @@
 <?php declare(strict_types = 1);
 
-namespace PHPStan\Analyser\Php;
+namespace PHPStan\Type\Php;
 
 use PhpParser\Node\Expr\FuncCall;
-use PHPStan\Analyser\TypeSpecifierContext;
 use PHPStan\Analyser\FunctionTypeSpecifyingExtension;
 use PHPStan\Analyser\Scope;
 use PHPStan\Analyser\SpecifiedTypes;
 use PHPStan\Analyser\TypeSpecifier;
 use PHPStan\Analyser\TypeSpecifierAwareExtension;
+use PHPStan\Analyser\TypeSpecifierContext;
 use PHPStan\Reflection\FunctionReflection;
-use PHPStan\Type\StringType;
+use PHPStan\Type\FloatType;
 
-class IsStringFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingExtension, TypeSpecifierAwareExtension
+class IsFloatFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingExtension, TypeSpecifierAwareExtension
 {
 
 	/**
@@ -22,7 +22,11 @@ class IsStringFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingE
 
 	public function isFunctionSupported(FunctionReflection $functionReflection, FuncCall $node, Scope $scope, TypeSpecifierContext $context): bool
 	{
-		return strtolower($functionReflection->getName()) === 'is_string'
+		return in_array(strtolower($functionReflection->getName()), [
+				'is_float',
+				'is_double',
+				'is_real',
+			], true)
 			&& isset($node->args[0])
 			&& !$context->null();
 	}
@@ -33,7 +37,7 @@ class IsStringFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingE
 			throw new \PHPStan\ShouldNotHappenException();
 		}
 
-		return $this->typeSpecifier->create($node->args[0]->value, new StringType(), $context);
+		return $this->typeSpecifier->create($node->args[0]->value, new FloatType(), $context);
 	}
 
 	public function setTypeSpecifier(TypeSpecifier $typeSpecifier): void
