+++ src/Type/Php/IsIntFunctionTypeSpecifyingExtension.php
@@ -1,18 +1,18 @@
 <?php declare(strict_types = 1);
 
-namespace PHPStan\Analyser\Php;
+namespace PHPStan\Type\Php;
 
 use PhpParser\Node\Expr\FuncCall;
-use PHPStan\Analyser\TypeSpecifierContext;
 use PHPStan\Analyser\FunctionTypeSpecifyingExtension;
 use PHPStan\Analyser\Scope;
 use PHPStan\Analyser\SpecifiedTypes;
 use PHPStan\Analyser\TypeSpecifier;
 use PHPStan\Analyser\TypeSpecifierAwareExtension;
+use PHPStan\Analyser\TypeSpecifierContext;
 use PHPStan\Reflection\FunctionReflection;
-use PHPStan\Type\FloatType;
+use PHPStan\Type\IntegerType;
 
-class IsFloatFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingExtension, TypeSpecifierAwareExtension
+class IsIntFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingExtension, TypeSpecifierAwareExtension
 {
 
 	/**
@@ -23,9 +23,9 @@ class IsFloatFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingEx
 	public function isFunctionSupported(FunctionReflection $functionReflection, FuncCall $node, Scope $scope, TypeSpecifierContext $context): bool
 	{
 		return in_array(strtolower($functionReflection->getName()), [
-				'is_float',
-				'is_double',
-				'is_real',
+				'is_int',
+				'is_integer',
+				'is_long',
 			], true)
 			&& isset($node->args[0])
 			&& !$context->null();
@@ -37,7 +37,7 @@ class IsFloatFunctionTypeSpecifyingExtension implements FunctionTypeSpecifyingEx
 			throw new \PHPStan\ShouldNotHappenException();
 		}
 
-		return $this->typeSpecifier->create($node->args[0]->value, new FloatType(), $context);
+		return $this->typeSpecifier->create($node->args[0]->value, new IntegerType(), $context);
 	}
 
 	public function setTypeSpecifier(TypeSpecifier $typeSpecifier): void
